+++ apps/bam.iobio/package.json
@@ -11,7 +11,7 @@
 	},
 	"dependencies": {
 		"clsx": "^2.1.1",
-		"components": "workspace:^",
+		"iobio-react-components": "workspace:^",
 		"react": "^18.3.1",
 		"react-dom": "^18.3.1"
 	},
