+++ package.json
@@ -1,5 +1,5 @@
 {
-	"name": "@overture-stack/iobio-react-components",
+	"name": "@overture-stack/iobio-components",
 	"version": "1.0.0",
 	"description": "Overture Integration of Iobio.io Genomic Analysis Tools",
 	"main": "index.js",
