+++ packages/iobio-react-components/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "components",
+	"name": "iobio-react-components",
 	"private": true,
 	"version": "0.0.0",
 	"type": "module",
