+++ packages/iobio-react-components/src/histogram.tsx
