+++ packages/iobio-react-components/src/percent_box.tsx
